Change: Generalize GitHubModel into a base Model class
The Model class can be used beside representing data from GitHub.
Therefore generalize it and move it to a new pontos.models module.
